report-generator/requirements.txt
datapane==0.14.0
pandas==1.4.2
matplotlib==3.5.2
pymongo==4.1.1
requests==2.28.0
PyYAML==6.0
jmespath==1.0.1
pytest==7.1.2
python-dateutil==2.8.2
google-cloud-storage==2.4.0
